+++ addons/yat/src/commands/builtin/Ls.cs
@@ -1,0 +1,130 @@
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+using System.Threading;
+using Godot;
+using YAT.Attributes;
+using YAT.Enums;
+using YAT.Helpers;
+using YAT.Interfaces;
+using static YAT.Scenes.Overlay.Components.Terminal.Terminal;
+
+namespace YAT.Commands
+{
+	[Command("ls", "Lists the contents of the current directory.", "[b]Usage[/b]: ls")]
+	[Threaded]
+	[Arguments("path:string")]
+	public sealed class Ls : ICommand
+	{
+		public YAT Yat { get; set; }
+		public Ls(YAT Yat) => this.Yat = Yat;
+
+		private CancellationToken _ct;
+
+		public CommandResult Execute(Dictionary<string, object> cArgs, CancellationToken ct, params string[] args)
+		{
+			string path = (string)cArgs["path"];
+			path = ProjectSettings.GlobalizePath(path);
+
+			_ct = ct;
+
+			return PrintDirectoryContents(path);
+		}
+
+		private CommandResult PrintDirectoryContents(string path)
+		{
+			if (!Directory.Exists(path))
+			{
+				Yat.Terminal.Print($"Directory '{path}' does not exist.", PrintType.Error);
+				return CommandResult.Failure;
+			}
+
+			try
+			{
+				string[] files = Directory.GetFiles(path);
+				string[] directories = Directory.GetDirectories(path);
+
+				DirectoryInfo directoryInfo = new(path);
+				FileSystemInfo[] fileSystemInfos = directoryInfo.GetFileSystemInfos();
+
+				StringBuilder sb = new();
+
+				AppendDetails(sb, fileSystemInfos);
+
+				Yat.Terminal.Print(sb.ToString());
+			}
+			catch (UnauthorizedAccessException)
+			{
+				Yat.Terminal.Print($"Access to '{path}' is denied.", PrintType.Error);
+				return CommandResult.Failure;
+			}
+			catch (PathTooLongException)
+			{
+				Yat.Terminal.Print($"Path '{path}' is too long.", PrintType.Error);
+				return CommandResult.Failure;
+			}
+			catch (Exception ex)
+			{
+				Yat.Terminal.Print($"Error accessing directory '{path}': {ex.Message}", PrintType.Error);
+				return CommandResult.Failure;
+			}
+
+			return CommandResult.Success;
+		}
+
+		/// <summary>
+		/// Appends details of the given FileSystemInfo array to the provided StringBuilder.
+		/// </summary>
+		/// <param name="sb">The StringBuilder to append the details to.</param>
+		/// <param name="infos">The array of FileSystemInfo objects to retrieve details from.</param>
+		private void AppendDetails(StringBuilder sb, FileSystemInfo[] infos)
+		{
+			int maxFileSizeLength = 0;
+			int maxLastWriteTimeLength = 0;
+
+			// Get the maximum length of the file size and last write time strings.
+			foreach (FileSystemInfo info in infos)
+			{
+				if (_ct.IsCancellationRequested) return;
+
+				maxLastWriteTimeLength = Math.Max(
+										maxLastWriteTimeLength,
+										info.LastWriteTime.ToString("yyyy-MM-dd HH:mm:ss").Length
+				);
+				maxFileSizeLength = Math.Max(maxFileSizeLength, (info is FileInfo file)
+									? NumericHelper.FileSizeToString(file.Length).Length
+									: 0
+				);
+			}
+
+			maxFileSizeLength += 4;
+
+			// Append the details of each FileSystemInfo object to the StringBuilder.
+			foreach (FileSystemInfo info in infos)
+			{
+				if (_ct.IsCancellationRequested) return;
+
+				var fileSizeString = NumericHelper.FileSizeToString(
+									info is FileInfo
+									? ((FileInfo)info).Length
+									: 0
+				);
+
+				var line = string.Format(
+					"{0}\t\t{1}\t\t{2}{3}{4}",
+					info.LastWriteTime.ToString("yyyy-MM-dd HH:mm:ss").PadRight(maxLastWriteTimeLength),
+					info is FileInfo
+						? fileSizeString.PadRight(Mathf.Clamp(maxFileSizeLength - fileSizeString.Length, 0, maxFileSizeLength))
+						: string.Empty.PadRight(maxFileSizeLength),
+					info.Name,
+					info is DirectoryInfo ? '/' : string.Empty,
+					(info.Attributes & FileAttributes.Hidden) != 0 ? "*" : string.Empty
+				);
+
+				sb.Append(line);
+				sb.AppendLine();
+			}
+		}
+	}
+}