+++ CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file.
 
 -   Ping command.
 -   Ip command.
+-   Ls command.
+-   FileSizeToString numeric helper method.
+
+### Changed
+
+-   List command alias renamed to 'lc'.
 
 ## [1.10.0-beta]